+++ app/src/main/java/com/steinwurf/petro/AudioExtractorActivity.java
@@ -5,25 +5,9 @@
 
 package com.steinwurf.petro;
 
-import android.Manifest;
-import android.app.Activity;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.media.MediaCodec;
-import android.os.Environment;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
+import android.support.v7.app.AppCompatActivity;
 
 public class AudioExtractorActivity extends AppCompatActivity
 {