+++ mediaplayer/src/main/java/com/steinwurf/mediaplayer/SampleProvider.java
@@ -1,0 +1,8 @@
+package com.steinwurf.mediaplayer;
+
+public interface SampleProvider {
+
+    boolean hasSample();
+
+    Sample getSample();
+}