+++ mediaplayer/src/main/java/com/steinwurf/mediaplayer/SampleStorage.java
@@ -4,33 +4,9 @@ import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
-public class SampleStorage
+public class SampleStorage implements SampleProvider
 {
     private static final String TAG = "SampleStorage";
-
-    class Sample
-    {
-        /**
-         * The timestamp in microseconds
-         */
-        final long timestamp;
-
-        /**
-         * The data buffer
-         */
-        final byte[] data;
-
-        /**
-         * Constructs a sample
-         * @param timestamp timestamp in microseconds
-         * @param data data buffer
-         */
-        Sample(long timestamp, byte[] data)
-        {
-            this.timestamp = timestamp;
-            this.data = data;
-        }
-    }
 
     private final List<Sample> samples = Collections.synchronizedList(new LinkedList<Sample>());
 
@@ -72,17 +48,19 @@ public class SampleStorage
      * Returns the number of samples in the storage
      * @return the number of samples in the storage
      */
-    public int getCount()
+    @Override
+    public boolean hasSample()
     {
-        return samples.size();
+        return samples.size() != 0;
     }
 
     /**
      * Returns the next {@link Sample}
      * @return the next {@link Sample}
-     * @throws IndexOutOfBoundsException if count < 1.
+     * @throws IndexOutOfBoundsException if no more samples are available.
      */
-    Sample getNextSample() throws IndexOutOfBoundsException
+    @Override
+    public Sample getSample() throws IndexOutOfBoundsException
     {
         Sample sample = samples.get(0);
         samples.remove(0);